Move instrument and market data logic into services directory
