+++ core/src/api/handlers.rs
@@ -1,11 +1,12 @@
 use crate::{
-    services::portfolio::get_portfolio_overview,
-    util::{
+    services::{
         constants::{OUT_DIR, SESSION_TOKEN_KEY},
-        db_helpers::{get_dividends, get_performance_signals, get_positions},
-        general_helpers::{get_env_variable, parse_timestamp},
-        taxation_helpers::get_events,
+        env::get_env_variable,
+        events::get_events,
+        parsers::parse_timestamp,
+        portfolio::get_portfolio_overview,
     },
+    util::db_helpers::{get_dividends, get_performance_signals, get_positions},
 };
 use axum::{
     extract::{Json, Query, Request},
