+++ core/src/cli/mod.rs
@@ -2,4 +2,5 @@ pub mod housekeeping;
 pub mod import;
 pub mod performance;
 pub mod portfolio;
+pub mod shared;
 pub mod taxation;
