+++ core/src/services/importers/revolut.rs
@@ -6,13 +6,15 @@ use chrono::prelude::*;
 use csv::{ReaderBuilder, StringRecord};
 use serde::Deserialize;
 
-use crate::util::{
-    db_helpers::{
+use crate::{
+    services::{
+        instruments::{identifiers::get_changed_identifier, ticker_symbols::get_isin_from_symbol},
+        parsers::parse_timestamp,
+    },
+    util::db_helpers::{
         add_dividend_to_db, add_fx_conversion_to_db, add_trade_to_db, get_listing_changes,
         Dividend, FxConversion, Trade,
     },
-    general_helpers::parse_timestamp,
-    market_data_helpers::{get_changed_identifier, get_isin_from_symbol},
 };
 
 #[derive(Debug, Deserialize)]
