+++ core/src/services/importers/wise.rs
@@ -6,13 +6,12 @@ use chrono::prelude::*;
 use csv::{ReaderBuilder, StringRecord};
 use serde::Deserialize;
 
-use crate::util::{
-    db_helpers::{
+use crate::{
+    services::{market_data::fx_rates::convert_amount, parsers::parse_timestamp},
+    util::db_helpers::{
         add_fx_conversion_to_db, add_interest_to_db, add_trade_to_db, FxConversion,
         InterestPayment, Trade,
     },
-    general_helpers::parse_timestamp,
-    market_data_helpers::convert_amount,
 };
 
 #[derive(Debug, Deserialize)]
