+++ core/src/services/instruments/identifiers.rs
@@ -1,0 +1,12 @@
+use crate::util::db_helpers::ListingChange;
+
+pub fn get_changed_identifier(identifier: &str, listing_changes: Vec<ListingChange>) -> String {
+    let relevant_changes = listing_changes
+        .iter()
+        .find(|item| item.from_identifier == *identifier);
+
+    match relevant_changes {
+        Some(listing_change) => (*listing_change).clone().to_identifier,
+        None => identifier.to_string(),
+    }
+}
