+++ core/src/services/instruments/mod.rs
@@ -1,1 +1,5 @@
+pub mod fund_data;
+pub mod identifiers;
+pub mod names;
 pub mod stock_splits;
+pub mod ticker_symbols;
