+++ core/src/services/market_data/mod.rs
@@ -1,0 +1,3 @@
+pub mod fred;
+pub mod fx_rates;
+pub mod prices;
