+++ core/src/services/market_data/prices.rs
@@ -1,0 +1,11 @@
+use rust_decimal::Decimal;
+
+use crate::util::db_helpers::get_instrument_by_id;
+
+pub async fn get_current_instrument_price(isin: &str) -> anyhow::Result<Decimal> {
+    let instrument = get_instrument_by_id(isin).await?;
+    match instrument {
+        Some(_) => Ok(instrument.unwrap().price),
+        None => panic!("No price found for ISIN {} in instrument table.", isin),
+    }
+}
