+++ core/src/services/mod.rs
@@ -1,6 +1,12 @@
+pub mod constants;
+pub mod env;
+pub mod events;
+pub mod files;
 pub mod importers;
 pub mod instruments;
+pub mod market_data;
 pub mod parsers;
 pub mod performance;
 pub mod portfolio;
+pub mod shared;
 pub mod taxation;
