+++ web/src/components/Charts/PortfolioPerformanceChart.tsx
