+++ web/src/components/Charts/PositionPerformanceChart.tsx
